added a basic projectile spawner with a projectile spawner prefab and a config for projectiles from the spawner
added a file to mechanics (GlitchEscape>Scripts>Mechanics>ProjectileSpawner) that contains the scripts for a projectile spawner and a projectile spawner projectile and its config. The prefab for the projectile is in Prefabs>Enemies>Attacks, the config for it is in Config>HDB, and there is also a prefab for the Projectile spawner in Prefabs>Enemies.