fixed a bug/added functionality to the projectile class
the projectiles would not delete themselves if they hit something other than an agent. added another check on collision within the projectile.cs code that deletes if hitting something that isnt an agent and after a brief period from being spawned, as the projectiles would delete themselves if any part of the projectile spawns inside of the original object it spawned from.