implemented fragment pickups
still tweaking minor things and adding some extra functionality, but should work for now. fragment maxes should be changed from a default value of 10 max to getting all fragments in the level and dynamically assigning fragments
FragmentComponent is in player and tracked fragment pickup total and FragmentInteraction tracks the actual pickup and deletion of fragment objects in the scene.