fixed player jump
Changed player jump to more traditional double jump. Player will only have one mid-air jump now, but jumping starting on a platform is another extra jump.
Raycast height decreased so the raycast is somewhat more accurate.
Added code IgnoreRaycast.cs for attaching to in-game assets that shouldn't be raycasted for the jump reset. Previously, the player's raycast would reset on any in-game object, like save points, text triggers, etc. It allowed the player to effectively jump infinitely when jumping over an invisible text trigger.
Alernatively, my fix to the raycasting could be changed to making the raycast ignore anything that isnt a platform. But this would still require changing all the platforms in the scene instead of all non-raycast objects. Not sure if that change is necessary or not.