Added jump debugging and PC debug controls
Jump has debugging rays for each time the jump raycast hits a wall or floor. However, debug.draw is super annoying and only shows in scene view with gizmos enabled, or in game view with gizmos enabled, but that makes it hard to play or see what's happening when in game view.
PC controlls are "`" or "~" to open the debug menu and the arrow keys, right, up, and down to navigate.