making sure to always call recalc positions at startup seems to fix this??? - would spawn otherwise but in the wrong location (at origin??) - might be a bug w/ initial spawning or something - fixed when fragments are selected (ie. editor script kicks in, which will sometimes call RecalcPositions() when stuff changes) - ergo this fix, which seems to work ¯\_(ツ)_/¯
note: issue is due to maintaining a list of game objects (which SHOULD
be serialized but might not), so we have fallbacks to:
1) get spawned objects from children + use those if the list is empty
2) if list is STILL empty (ie. objects were not spawned), we can
respawn (recalc will call respawn if it's empty)
also 3) can't actually fully respawn particles (or change particle
count) b/c child objects under a prefab are effectively locked and can't
be deleted. fortunately, no prefabs seem to need to actually do this, but
if you try (eg. change # particles at runtime on a prefab) you will ofc
get a bunch of console errors (fwiw)

long term solution is to switch to unity ECS / DOTS which would obviate
all of this (and might be more performant), but it's doubtful whether
we'll have time to do that :/