identified + fixed bug / missing feature where inactive fragments (ie. under an inactive glitch maze, etc), were not detected by the SceneFragmentManager custom editor, and did not have their IDs assigned.
This was actually somewhat complicated as there isn't any builtin
function to get all scene objects, including inactive ones. Those are
skipped by GameObject.FindObjectsOfType<T>()), and
Resources.FindObjectsOfTypeAll<T>() finds those objects but also finds
a LOT of other stuff as well

The fix was to write a new utility function that did this, and while
this took a bit of digging (and debugging) this now works, and I've
proceeded to assign all missing fragment types.

Also, as a weird accidental bug / bugfix, platform fadeout now... works.
I have no idea why, but I'm happy that it works now, and yes, visually
it seems to work pretty well...

(yay for having that finally actually work and on a branch that I'm
working on, lol)